moved board to page, timer set up is now on the board
